Fix Windows issues

+++ src/Framebuffer.cpp
@@ -78,7 +78,8 @@ void Framebuffer::bindColorTexture(const ShaderProgram *shader) const
 	glUniform1i(shader->getLocation("headerTexture"), 1);
 	glActiveTexture(GL_TEXTURE1);
 	glBindTexture(GL_TEXTURE_2D, headerTexture_);
-	glUniform2f(shader->getLocation("frameSize"), size_.width, size_.height);
+	glUniform2f(shader->getLocation("frameSize"),
+		static_cast<GLfloat>(size_.width), static_cast<GLfloat>(size_.height));
 }
 
 
@@ -87,5 +88,6 @@ void Framebuffer::bindDepthTexture(const ShaderProgram *shader) const
 	glUniform1i(shader->getLocation("depthTexture"), 0);
 	glActiveTexture(GL_TEXTURE0);
 	glBindTexture(GL_TEXTURE_2D, depthTexture_);
-	glUniform2f(shader->getLocation("frameSize"), size_.width, size_.height);
+	glUniform2f(shader->getLocation("frameSize"),
+		static_cast<GLfloat>(size_.width), static_cast<GLfloat>(size_.height));
 }
